+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 ------------------
 
 * Laure validation events on Slack use the complete_feedback field, not feedback (ericof).
+* Professional: On approval send slack message to Finance (ericof).
 
 2.0.6 (2017-02-21)
 ------------------